MessageBox: don't show scrollbar if not necessary
Sometimes a scrollbar was shown even for small message boxes, especially
if the text size of windows is set to medium (125%). This happened only
with Chrome, not with IE and FF.

Solution is to use the exact size in the MessageBoxLayout to set the
height of the content instead of a rounded value.

209700