DesktopTable: border is drawn wrongly when zoomed
When the page is zoomed or windows font size is set to medium (125%),
the 2px border sometimes has a white gap between the cell border and the
row border which looks bad. This seems to be a miscalculation by the
browser. 

To prevent that the border is reduced to the simple 1px border.