Table: header menu is positioned wrongly when page is zoomed
When the page is zoomed or windows font size is set to medium (125%),
the header menu is positioned too far down which reveals a grey line
that is not supposed to be visible.

Fixed by using the exact bounds. Also removed the confusing
includePosMargin argument from scout.graphics.bounds and
scout.graphics.offsetBounds. These methods now take an options argument
which will be passed to the scout.graphics.size function.