SmartField2: prevent too many requests, removed duplicate code
Problem was that too many requests have been sent to the sever when the
popup was closed and the user typed fast. Because for the openPopup case
no debouncing was done. Now a flag prevents multiple execution of
openPopup and the lookup performed there uses the same debounce code as
all other lookups.