DatePicker: wrong month width when getting invisible
When an editable date field with an open picker is scrolled out of view,
the picker will be made invisible because the field is not visible
anymore. If that is the case, the month width may not be determined
correctly -> use visibility: hidden instead of display: none to make it
invisible.