SmartField2: fixed bug with wrong focus position after cell-edit
After cell-editor-popup has been removed, the focus was at the wrong
position. The reason was that _onFieldMouseDown called focus() on the
editor field. Because the focus-context for the cell-editor-popup was
not installed at that point in time, the field was stored as last valid
focus position on the focus context of the _desktop_ which is wrong.

The bugfix simply calls togglePopup, because the enabled check is not
required for the cell-editor case