Scrollbar: Correctly remove attached scroll listeners from anchestors
The "anchestor scroll or resize handler" ist attached to all anchestor
DIVs of the scrollbar. When the scrollbar is removed, those listeners
are detached again. Previously, the same jQuery selector was used in
_render() and _remove() to find the anchestor DIVs. However, if the DOM
has changed between those two calls, not all listeners might have been
removed correctly. This leads to dangling listeners that can cause JS
errors later.

Example: Scrollbars inside a HTML field are removed from the DOM when
the field's content is replaced (because the $.fn.html() method is
used). When the scrollbars are finally removed using the Widget.remove()
method, the anchestor DIVs can no longer be resolved correctly.

To fix the issue, the list of anchestors is remembered in _render() and
is then used in _remove() to remove the listeners from exactly the same
anchestors (similar to $parent that is also stored in a variable).

211599

Change-Id: If3c6c536ef4fb752877e44db605e6f2e8dc37f83