Pointer cursor not showing anymore on link buttons (IE)
When hovering over an enabled link button, the cursor should be a
pointer (hand). With IE 11, this does not work if the element has
display: inline and pointer-events: none (pointer-events was added with
commit a63788016f6a26a71a3f54d7f5ac43806e0462ac).

Fix is to set pointer-events to auto which prevents having an active
state but since the link-buttons don't have an active state by default
it is fine.

211641