Improve submenu icon and fix link button cursor (IE)
When hovering over an enabled link button, the cursor should be a
pointer (hand). With IE 11, this does not work if the element has
display: inline and pointer-events: none (pointer-events was added with
commit a63788016f6a26a71a3f54d7f5ac43806e0462ac).

Fix is to use display: inline-block instead of inline. To make vertical
alignment work again the vertical-align-helper is used. Since the
submenu icon is too big the button is now misaligned if there is a
submenu. To properly fix that and to make the submenu icon look good in
all browsers, the icon is now set by JS instead of CSS. I did not found
any hack free CSS only solution which works in every browser.

211641