graphics: Allow a boolean as options argument (short-hand)
When only setting the "includeMargin" flag for a graphics method call,
it is quite cumbersome to pass an entire options object. As a
convenience, a short-hand version of the methods is now supported. When
passing a boolean instead of an object, that boolean is automatically
converted to an options object containing the "includeMargin" option.

Change-Id: I41bb645efca3aed17a67117980b3f093a121cce2