FormMenuPopup: don't close on resize events
Instead layout and position which is especially important for mobile
devices if the popup contains an input field.
In that case activating the field opens the keyboard which may resize
the screen..

Observed on android tablets. Windows tablets don't resize but seem to
trigger a resize event anyway which will close the popup. With iOS it
would work fine because it does not resize when the keyboard opens.