Tree: node width is wrong if tree is invisible initially
Commit f36c955b8e2dd3b03c110f151cd4112f9c435529 changed the time when
the tree is rendered. Previously, the tree was rendered and afterwards
the visible flag applied. After the change the tree was rendered after
the visible flag was applied which had the effect that nodeWidth was
wrong.

The commit is actually a workaround, to completely fix it we intend to
change the behavior of renderVisible so that only visible widgets are
rendered.