Form: fix modal handling
Modal dialogs did not work so far for Scout JS dialogs (nor did it
anymore for Scout classic since the last commit).

Reason is the confusing display parent handling. Refactored so that
display parent is set when form is opened and removed confusing
glass pane logic in form.init. Also made sure that parent always points
to the display parent, if there is one (important for scout JS case).