DateField: allow dynamic toggling of hasDate and hasTime
When toggling hasDate and a date was entered, the date should be
preserved so that when toggling again the previous date is shown instead
of 1970-01-01. But: When no date was entered and hasDate is false
(relevant for time only fields), 1970-01-01 should be used instead of
the current date to not break existing applications which may rely on
that behavior.

Also contains a bugfix in _predictTime: date had been reset to reference
date if time was predicted.

212505