DesktopToolBox: invalidate layout when menu images load
The layout of the menu box in the desktop header needs to be invalidated
after an image loads, otherwise it is likely that the menu won't have
the correct size.

Introduced scout.Icon which may handle font and image icons and makes
use of scout.Image for the loading behavior (renamed existing scout.Icon
to scout.IconDesc).

The same is not possible for the menubar because layouting rebuilds
every item which would lead to a loop

212286