SmartField2: fixed bugs with smart-field as cell editor.
The initial value of the smart-field was never set correctly, only
displayText and lookupRow was sent to the browser. This worked somehow
in normal smart-fields but caused errors in the cell editor, for
instance it was not possible to delete (empty) a smart-field in a table.

In the touch-popup we now pass the _embedded_ field, instead of the
_touch_ field. That way the embedded field executes lookups, which
simplifies things.