TableControl: wrong content after fast clicks on different controls
Use case: double click on a table control and while the container is
still closing click on another control. The result is that the content
of the first control is displayed along with the content of the second
control.

Reason: the done callback of the close animation has never been executed
due to calling stop(true).

Solution: remove the content of the currently rendered control if the
container is opened again while closing and the current control is a
different one.

178419