AggregateTableControl: fix exception when column gets invisible
Use case: table has a number column and a string column. Aggregate
control is open. If the string column is removed, all rows will be
deleted and newly added due to the column structure change. The
aggregate control will be notified but instead of receiving the
all_rows_deleted event it receives the rows_selected event first. This
will result in an error because table.filteredRows is not updated yet
but table.columns is -> Accessing the cells is not possible because
filteredRows.cells still has a size of 2 but columns has a size of 1.

Also fixed duplicate rendering of aggregate cells. RenderContent called
_aggregate and _renderAggregate and both added the aggregate cells to
the container. Reason: rendered belongs to the control and not to the
content, the control is always rendered when _renderAggregate is called.

212882