IE: layout breaks when toggling tree node check box in a tab box
Use case: tab box which contains a tree box. Toggling a check box from
checked to unchecked of that tree will invalidate the tree box layout
because the tree box property 'marked' changed. 

Error: this will break the layout completely because IE 11 suddenly
returns 0 as width when calling getBoundingClientRect() for the desktop
container. When triggering the layout anew (e.g. by resizing the browser
window), everything will be fine again.

Reason: it is not comprehensible to me why it happens. But I noticed
that it works if the content of the before element is not explicitly set
to '' when the checkbox is being unchecked. Maybe IE has troubles
removing the content of a focused element?

212842