RadioButton: remove before pseudo element
IE breaks the layout if a radio button is toggled and a tree box layout
invalidated. Popular example is the active/inactive radio button group
of a tree box (see widgets examples).

This is a similar issue to 71799a956fe2e0342f0aaee05ce3c94faf6935fa. To
avoid such problems in the future and to clean up some other workarounds
around that before element, the radio button has been refactored so that
the circle is not drawn by a before pseudo element anymore. It is now
implemented the same way as a checkbox.

Also removed ButtonLayout because $fieldContainer and $field is the
same. Added RadioButtonLayout instead because it is not possible to set
the max-width using CSS due to dynamic icon width.