IE: layout breaks when selecting radio button of tree box
IE breaks the layout if a radio button is toggled and a tree box layout
invalidated. It suddenly returns 0 for treeBox.$container[0].offsetWidth
and getBoundingClientRect().width. ClientWidth returns the correct
value. 

Popular example is the active/inactive radio button group
of a tree box (see widgets examples).

This is a similar issue to 71799a956fe2e0342f0aaee05ce3c94faf6935fa. It
looks like IE has trouble toggling the pseudo element when it is
clicked. Solution is to always add content to the after element and only
toggle its visibility. This solution is different to the one in 7.0
where the radio button has been refactored to not use a pseudo element
for the circle anymore.

212842