Fix request timeout
The await timeout for model jobs to complete is currently set to 1 hour.
After that hour, a TimedOutError occurs. Since it is an error it will be
sent as internal server error 500 to the client. The long running model
job is still running, so the next click after the timeout cannot be
processed.

Expected behavior: The error should not be sent as internal server
error, instead a message box should be displayed saying a request
timeout has occurred. Also, the running model job has to be aborted so
that the user can still use the ui after a timeout.

Added property called scout.ui.model.jobs.await.timeout to make the
timeout configurable.

Also made sure that PlatformError does not fall through the error
handler in JsonMessageRequestHandler, instead a json error response
should be sent to the client.

209789