Fixed problem with flickering scrollbars and table-settings icon.
IE seems to have problems with rendering static positioned elements
inside a DIV that has rounded corners -> scrollbars flickered wildly
while they have been moved around. 

We noticed earlier that with rounded corners also the performance of the
table suffers, so this bug is also solved with this commit. Of course
the commit has an impact on styling, as some fields still have rounded
corners and others don't. But we think that performance and flicker-free
scrollbars are more important.

The change in TableHeader.less addresses a Bug on certain zoom levels in
IE where the table-settings icon (cog) lapped into the border-border of
the table-header, which could only be noticed in the 'dark' theme where
the bottom-border is 2px. Also removed the fix/unfix code because it's
not required anymore.

180466 176222