MenuBar: Set menu as default menu if the flag "defaultMenu" is set
Previously, the menu bar checked only the menu's key stroke to determine
the default menu. The flag "defaultMenu" in Menu.js can now be
explicitly set to true or false to override this behavior for a specific
menu (independently from the key stroke). By default, the key stroke is
still checked.

Change-Id: Ib8833556a2f43c3eb2f065af19f017a0a2d5b1fe