Table: enable sort click without modifier if header menu is disabled
If the header menu is enabled, clicking the header opens the menu.
Ctrl-clicking the header sorts the column. If the header is disabled, it
should not be necessary to press ctrl, it is more intuitive to just
click it. This also allows sorting on mobile devices where it is not
possible to press ctrl.