Prevent waking up the poller unnecessarily on "timed out" errors
When the model job to display the "request timed out" error is executed,
the m_processingJsonRequest flag is already reset, even though the user
request technically is still being processed. This causes the poller to
be signaled unnecessarily. To let the "UI data available listener" to
ignore this job, a special execution was added.

209789

Change-Id: I40012a2cd94ef81ec0875b508465d408f1907371
Reviewed-on: https://git.eclipse.org/r/103454
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>