Table: fix flickering of table settings icon
This partially reverts 1179821b07b2ae6e59db55d3917bab3e67276f39.

Using scrollbars.fix is necessary to prevent flickering. The problem of
that function is, that it uses scout.graphics.offsetBounds which has 2
drawbacks:
1. It is not exact. This is the reason why the settings icon sometimes
is too large.
2. It uses jquery.offset() which is different to
getBoundingClientRect().top because it adds pageYOffset which in case of
IE and 125% zoom is 1. Reason: the desktop height may be fractional
(e.g. 677.6) but the navigation height not because our layout always
rounds the values up (see scout.graphics.getSize).