Widget: Resolve references in model when creating widget instances
The value of so-called "widget properties" are other widget instances.
When creating a widget from a model, the models of those other widgets
may directly be specified as nested objects. However, this approach does
not work if a widget instance is referenced in multiple widget
properties (e.g. a list of "items" and one single "selectedItem"). In
the remote case (model adapters), it is possible to simply pass the
referenced adapter's ID instead of the adapter data itself. The same
should be possible for ScoutJS (UI-only) widgets.

Because in ScoutJS there is no central widget registry, the passed IDs
can only be referenced locally (using the Widget.widget() function).
This is enough for simpler cases. If the data structures are more
complicated, the widgets have to be linked manually after creating them.

Change-Id: I9196016dae666146f37badd1142f353b0e843731