Session: Ensure poller does not permanently stop on JS error
When a JS error occurs during processing of a response to a ?poll
request, the "backgroundJobPollingSupport" state must not remain
RUNNING. Otherwise, the poller will never be rescheduled again, because
Session.js assumes that the poller is still running. This may cause the
server session to be invalidated unexpectedly (because there are no
?poll requests keeping it alive).

To fix this, the AJAX callback functions must ensure that the
"backgroundJobPollingSupport" state remains _only_ RUNNING iff the poll
request is rescheduled again. In all other cases the state must be set
to FAILURE or STOPPED.

Change-Id: I9b9b81d209c44ffe9a0b93706e5f58f3bb543e38
Reviewed-on: https://git.eclipse.org/r/103677
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/103688
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>