Table.js: add buffer for smart columns
Smart columns use promises to lookup the text for the value. Since
jQuery 3 every promise is resolved asynchronously which means the text
lookup will resolve always when the row has already been added and
rendered. This will generate a lot of updateRow calls and therefore a
lot of renderViewPort calls which are quite expensive.

Solution is to buffer the update and call updateRows only after all
promises have been resolved.

213394