SmartField2: improved handling of errors and warnings
Temporary warnings like 'no data' must not replace already existing
errors like 'no results found'. For this we now have the property
'lookupStatus' which is used to display the status of the FormField
while lookups are performed. When we leave the field, the lookupStatus
is discarded or set as real error, depending on the error code.