SmartField2: fixed problem in multiline smartfield
Multiline smart-field displayed a 'not unique' error for no reason, when
the first line of a proposal existed multiple times in a lookup result.
However, in a multiline smart-field it is not enough to look at the
first line, we must also consider the additional text lines.

Addes some selenium utils used in the test-suite.