DesktopBench: fix splitter position in bench
"diff" is the amount of pixels the splitter should be moved. This value
did not include the padding. After the calculation of the splitter
center in Splitter.js was corrected (commit
fc924c1c29d8e499d3a3159888ed0359a82d16dc), the resulted in "jumping"
splitters in the desktop bench.

Change-Id: I0681e6ca3d298757d68e6dea4f203f60d00f5f69