Table: allow the configuration of the column minimum width
If a regular column should be made very small (e.g. 32px), this was not
possible because the ui defined a minimum width of 60px. Only boolean
and icon columns were allowed to be smaller than that.

Because there are cases where such small columns may be useful, the
configuration of the minimum width is now possible using the new column
property minWidth.

This was mainly added to make the organize column form smaller.

213542

Change-Id: I2516c1b4c7b0e39c30cd97a6cf808a4e2e5cd9d5
Reviewed-on: https://git.eclipse.org/r/104110
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/104175