DateField: fix time picker in touch mode with editable tables
If an editable time column is used with a touch device and the user
selects a time using the picker, the table row is not updated with the
entered time.

Also, when the date or time touch popup closes, the cell editor should
close as well. There is no need to make the user click a second time
after he chose a date.

213807