DateTimeField: clear should not clear date AND time
When pressing the clear icon of a datetime field, only the content of
the field with the clear icon should be cleared. E.g. pressing the "x"
in the time field, the date should not be removed.

Also: when switching between date and time field, the content should be
preserved. E.g. Focus is in the date field, content is removed and tab
is pressed so that the time field is now focused. The content of the
time field should still be there.

Also: when leaving the datetime field and one field is empty and the
other has invalid content, an error
should be shown rather than removing the content of the other field.
E.g. Invalid date is entered, time field is focused. Time is cleared and
tab
pressed -> Because neither date nor time field is focused acceptInput is
triggered which will set an error status because date has an error.

213466

Change-Id: Ic3baf1ac522ff3de7a4f28bb4fdf52948f17b5a1