MenuBar: prevent flickering when changing visibility
If the visibility of a menu is changed using java, the menu bar might
flicker because the ui updates the menu bar as soon as possible, but the
visibility change from the server is processed asynchronously.

Example: There is no row selected yet, single selection menus are not
visible. The user selects a row and the ui shows the single selection
menus. If the java code now makes some single selection menus invisible
based on the selected rows, the menus will disappear again but were
visible for a blink of an eye.

Solution is to wait for pending requests before updating the menu bar.
Also made sure that the layout is validated after the request callback
is executed, otherwise the menubar might still flicker. 

Example: no row is selected, user selects a row and the server will make the empty space menus invisible. This will eventually hide the menu bar because updateMenuBar has not been executed yet. Now single selection menus are made visible, as well as the menubar which will invalidate the table layout. But the menubar still has a height of 0 because table layout has not been validated yet -> Flickering because the next validation will be done using setTimeout.