DateField: partially revert 'clear should not clear date AND time'
The commit broke an essential behavior: filling in a value in the date
field and pressing tab should complete the prediction for the date and
suggest a time.

Reverted the part of the change which prevented the removal of the time
when a date was deleted but the time field focused. This is actually not
really necessary, it may be even annoying because the user has to delete
the time separately.