Use class 'hidden' for setVisible
JQuery's show() and hide() has several problems:
1. They are not very fast because hide() has to read the current display
style, so that show() can revert to that style.
2. Show() will set the style explicitly which may override existing CSS
styles (e.g. CSS sets display: none, show will set it to display: block)

It is a lot easier to set a class which will set 'display: none' if the
element should be hidden.

show() and hide() are left unchanged.