Table: Prevent shifting text when clicking to edit a clearable cell.
This prevents the content in BasicFields within TableFields from moving
a few pixels to the side when the clearable icon is shown, for cells
that are left- or right-aligned. The behavior is not changed for
centered cells, as the additional left padding needed would result in a
large amount of whitespace and a smaller editing area.

213787

Change-Id: I9a407a3d31f2b4387a6528a042b0751d7b4d28b8
Signed-off-by: Jeremie Miserez <jeremie.miserez@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/105077
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>