Use compact style for fields with width <=61 instead of <50
This enables the compact style for TableField cells with a minimum
configured column width of 60. As the first column may be 1px wider due
to a missing 1px border, 61 is chosen as the minimum width value. On
fields and TableField cells that are clearable (X icon), this leaves
more room for the content and prevents accidentally clearing the field
by clicking on the clearable icon.

Change-Id: If4bb405d3998eeba19cd0cbe3b46a3bc2a5795b7
Signed-off-by: Jeremie Miserez <jeremie.miserez@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/105087
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>