Fix potential JS error when opening a ContextMenuPopup
Opening of a context menu popup is asynchronous in most widgets, because
the server-side model could alter the menu items to show. Because of the
asynchronous nature of this function, the container widget could have
been removed or detached in the mean time. Opening a popup on a detached
or removed component results in JS errors. To prevent such errors, a
generic check as added at all asynchronous functions that create a
ContextMenuPopup.

214426

Change-Id: I7e8bbbd49df60dc09c462501902cd7c725b746ac
Reviewed-on: https://git.eclipse.org/r/105556
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>