Fix potential JS error when opening a ContextMenuPopup
Opening of a context menu popup is asynchronous in most widgets, because
the server-side model could alter the menu items to show. Because of the
asynchronous nature of this function, the container widget could have
been removed or detached in the mean time. Opening a popup on a detached
or removed component results in JS errors. To prevent such errors, a
generic check as added at all asynchronous functions that create a
ContextMenuPopup.

214426

Change-Id: If431a4eb50b0c55afe925cb083541a799f8d167e
Reviewed-on: https://git.eclipse.org/r/105556
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>