DefaultValues: improve overriding
A default values contributor with a lower order (e.g. 10) should
"override" a contributor with higher order (e.g. 100).

Example: 
The objectType hierarchy in the original contributor contains "Widget":
{"Tile": null}. 
A contributor with lower order contains "Widget": {"Tile":
{"SimpleTile": null}.

Until now, this did not work, SimpleTile was always removed.

To solve it the contributors have to be processed in reverse order.