SmartField2: cleanup various lookup results.
In previous version there were multiple result types for smart fields
lookups. These different types were superficial and have been removed
with this change. Additionally the RemoteLookupCall implements now the
lookupByKey method, which prevents an error when a value is set on the
smart-field that does not exist in the data source of the lookup call.

213886