SmartField2: removed parameter 'browseHint'
Removed parameter 'browseHint' from all "browseLookup" methods. The
parameter was not required and was set to various values like null, ""
or this.getWildcard(). The parameter was now removed entirely and the
smart field always passes getWildcard() to the lookup call. 

This means it's still Ok, to check if call.getAll() returns != null, but
you should not build logic around the value of getAll() contains an
expected string.

Also use the QueryBy enum in the JSON protocol and make the JS
implementation consistent with the Java implementation.