SmartField2: fixed error when an invalid value is set on the smart field
When smart field was set to a value that did not return any results in
the lookup call an error occurred in remote / Scout classic case.
Usually we don't need to do a lookupByKey in remote mode, because the
server sets either only a displayText or a lookup row which contains
value, displayText. However, when the server does not find a lookup row
for a given value it sends the value to the UI which starts a lookup
when it tries to render the display text for that value. That's why we
implemented lookupByKey() for the RemoteLookupCall too.

Note: the new impl. does not cause errors in the UI anymore. However, 
the display text will be empty, when the field is set to an invalid
value.

213886