Screen moves when busy indicator is shown and browser zoomed
With windows zoom set to 125% or browser zoom enabled, the desktop
height may be fractional. Since the layouts always round the sizes up,
the navigation and bench will be slightly bigger.

This leads to a problem when the glass pane is inserted. It will look as
the screen moves up and down about 1px.

215027