NumberField: reject invalid formulas (prevent "NaN" display text)
The Calculator's check to determine if an input is a "formula" consists
of a regular expression match. It is too simplistic to recognize invalid
expressions such as "1.2.3". Therefore, evalFormula() might still return
NaN. This result has to be caught and an error has to be thrown to make
the field invalid (error status). Otherwise, the text "NaN" would be
shown as display text.

214654

Change-Id: I7f84d296334f39cc3cf64651e0aed37a31c6318b
Reviewed-on: https://git.eclipse.org/r/106081
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>