getContribution() should find all nested contributions
The recursion in getContribution() failed if the contribution to find
was not part of the first branch of sub-contributions. This is
especially dangerous because the order in which the contributions are
put into the map is random and varies from application launch to
application launch (because hashCode of java.lang.Class is not
deterministic).

215087

Conflicts:
	org.eclipse.scout.rt.client/src/main/java/org/eclipse/scout/rt/client/ui/basic/table/columns/AbstractMixedSmartColumn.java

Change-Id: Ia2781b455a8af89eeb8a1eea65815b384087d235
Reviewed-on: https://git.eclipse.org/r/106138
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>