getContribution() should find all nested contributions
The recursion in getContribution() failed if the contribution to find
was not part of the first branch of sub-contributions. This is
especially dangerous because the order in which the contributions are
put into the map is random and varies from application launch to
application launch (because hashCode of java.lang.Class is not
deterministic).

215087