Tiles: remove mandatory indicator from tiles
The tiles field may show an indicator but the tiles itself should not
add a margin
(otherwise it would be hard to align form field tiles with non form
field tiles). 
If tiles should really show a mandatory state once, the
indicator should be drawn inside the tile, probably near the label.