Planner: menus are not shown when visibility changes
If the visibility of a resource menu is set to true in
execOwnerValueChanged, the menu is not shown after the user selects a
resource (by clicking on the resource name). Only after clicking a
second time the menu appears.

Reason: invisible menus are not added to the menu bar, so when the
property change event is processed the menu is not rendered and may
therefore not be made visible.

Fixed in the same way as it is done for table and tree.

214554