Fixed 'not unique' bug in tree smart fields
When user typed a proposal that should return an exact (unique) match,
the smart field displayed a 'not unique' error, because a hierarchical
lookup adds parent rows to the result. So our numLookupRows > 1 check
for uniqueness failed. However: the proper check was already implemented
in the _extendResult function, but was not called by the
_lookupByTextOrAllDone function.

Renamed property 'singleMatch' to 'uniqueMatch' to make naming more
consistent.

214155