Tree: fix exception with nested delete all events
Use case:
Tree with following structure:
A
|-B 
  |-C with Table

1. RemoveChildNodes for B -> allChildNodesDeleted event for B
2. RemoveChildNodes for A -> allChildNodesDeleted event for A

TreeEventBuffer will remove first event so JsonOutline only processes
the second one. It should now dispose the table for node C, but this
won't happen because B.getChildNodes().size() will return the actual
child node size which is 0 because they are removed from the model.

Solution: JsonOutline needs to keep track of the child/parent link so
that really all nodes are disposed on a delete event.

215123

Change-Id: Ibc8b0a9bbc6a6fb241f93a8e8787d0c136eb31f4
Reviewed-on: https://git.eclipse.org/r/106525
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/110374
Reviewed-on: https://git.eclipse.org/r/110375