Fixed layouting bug which caused JS errors in some cases
The error caused by this bug was a 'Already rendered' error for a menu
item in the MenuBar. The error occurred because the menu item was
removed in removeMenuItems, which caused the focus manager to choose a
new focusable item. This item was s StringField. On focus the VaueField
upates the clearable property which called revalidateLayoutTree(false).
This caused the layout-validator to call layout on the same Layout
instance again, which led to an invalid internal state in the
MenuBarLayout. This situation is now prevented by calling
revalidateLayout() instead. It's not required to invalidate parents.

215475