Tiles: support menus
Tiles now support context menus. They don't have a menu bar because
there is no need yet. 
Added a MenuMediator so that the menus of the tiles may easily be added
to another widget, like a group box. Also added a utility to update the
visibility according to the menu type.