Fix exception when pressing f5 while popup blocked message is shown
Use case: 
1. Open a form in a separate window
2. Popup blocked notification is shown
3. Reload the page

Already rendered exception occurs because notifications are rendered in
postRender of the desktop after the forms are rendered. If one form
opens a notification, the function renderNotifications will fail because
the notification is already there.

Solution: it is not correct to render the notifications in post render,
moved to _render.

215546