Fix exception when reloading while popup with glass pane is open
Use case: 
1. Open mobile mode
2. Open a search form
3. Reload the browser page -> exception

Reason: when the glass pane targets are collected during desktop
rendering, a DeferredGlassPaneTarget is created. After the desktop is
rendered, the deferred is resolved an the glass pane rendered. The
exception happens, because the DeferredGlassPaneTarget expects the
glassPaneTarget to be a jQuery object and therefore uses [0] to get the
HTMLElement. This will be undefined if the $glassPaneTarget itself
already is a HTMLElement.

215551