Table: prevent programmatically selecting filtered rows
When a selected row is filtered (= not visible anymore), it is
automatically deselected. If the filtered row is programmatically
selected again (e.g. by the server model) it should not be selected.
Otherwise, the behavior would be inconsistent (differed order of
operations would result in different results). Also, this would lead to
follow-up errors, e.g. when a "scroll to selection" event occurs. This
cannot be done, because the selected row would not be rendered.

214415

Reviewed-on: https://git.eclipse.org/r/109968
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>