Table/Tree: prevent programmatically selecting filtered rows/nodes
When a selected row is filtered (= not visible anymore), it is
automatically deselected. If the filtered row is programmatically
selected again (e.g. by the server model) it should not be selected.
Otherwise, the behavior would be inconsistent (differed order of
operations would result in different results). Also, this would lead to
follow-up erros, e.g. when a "scroll to selection" event occurs. This
cannot be done, because the selected row would not be rendered.

The same applies to Tree/TreeNodes.

214415

Change-Id: I13af8e25074336ab37720fd474bed00805f9e43c
Reviewed-on: https://git.eclipse.org/r/109968
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>