Calendar: ensure component to resolve is contained in PROP_COMPONENTS
Selected component must be part of the list of all components. This
check is necessary e.g. when deleting a component. The components are
reloaded, but the selected component is not updated automatically
(because it is a separate property). The resolveComponent() will now
return null when the given component is not contained in the list of all
components, thus clearing the selection.

215698

Change-Id: I691396013fc22b2892e532134502e360ebe3968c
Reviewed-on: https://git.eclipse.org/r/110540
Tested-by: Hudson CI
Reviewed-by: Hannes Mueller <hannes.mueller@bsi-software.com>