Table throws error when table is detached and row is removed
The consistency check may not be performed when table is not attached
and rendered. When the table is detached it can still add rows, but
these new rows are not rendered while the table is detached. Thus this
check would fail, when a row that has been added in detached state is
removed again while table is still detached.

216194