Don't select a view after page reload if no view was selected before
If no view is selected, reloading the page will select the previously
selected view. This is especially bad, if a modal dialog is open with
outline as display parent. The dialog won't be visible anymore after
page reload, because the desktop is now in background due to the open
view.