Form: kill form on server when it is closed by the UI
Normally, in Scout classic forms are closed by the model. This causes a
"formHide" event on the desktop and a "disposeAdapter" event on the
session. But when the form is closed by the UI, the widget was just
destroyed. It remained open in the server-side model, which could cause
errors when model events happened on that form (because the UI did not
know the adapter anymore).

This could be reproduced like this:
- Open a form with the display hint POPUP_WINDOW.
- Click the browser's native "x" button of the popup window.
- DesktopFormController.js registers that the popup was closed, and
  correctly calls form.close(). However, this just destroys the form
  widget, but does not notify the UI server (verify in the network
  inspector).
- Press Ctrl-R
- The page reloads and opens the previously closed form again.

This is fixed by sending the 'close' event to the server. The JsonForm
then kills the form. This is important, because the native popup window
is already closed. The model can therefore not cancel the form closing
or show any message boxes.

216643

Change-Id: I7378f4a9e725c7da4e46def755b35235a15605e5
Reviewed-on: https://git.eclipse.org/r/110825
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>