Fix logging of cancelled client jobs due to session shutdown
All client jobs assigned to a particular session are canceled when the
session is shutting down. The affected jobs query did not restrict the
job state, so that completed jobs might be returned as well (the
FutureSet that manages all running jobs references DONE jobs as well, at
least for a _small_ amount of time). Excluding DONE and REJECTED jobs
prevents logging of false-positives.

216696

Change-Id: I514f8c68e5a2015fa986a802601e290e6f6a3311
Reviewed-on: https://git.eclipse.org/r/110843
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>