Tiles: add insert animation
Insert and delete animation is disabled for placeholders because when
the gridColumnCount is reduced, the placeholders will be added and
removed unnecessarily which would be visible if they were animated.

Also added property startupAnimationEnabled which is currently only used
by the page tiles.

Also fixed exception when a placeholder tile was clicked (remote).