Fixed problem in Table.js #_renderViewport (attachedAndRendered)
The previous version used the Widget hierarchy to check if the widget
and all of its parent have been attached. But sometimes the widget and
the DOM hierarchy are not in-sync. Thus the previous version could
result in wrong results. The current version simply checks if the
$container of the widget exists in the DOM, because that's the most
reliable way to check if the DOM element is attached.

This solves a problem with smart-field popups in dialog-forms, that did
not open correctly when the outline was not visible at all (=minimize
tree).

215892