Add warning if the wrong thread is used for model updates
With commit f8f68a76fb3d84ea0be6b5a0712e6ca4396186b5 a property has been
introduced to throw an exception if the wrong thread is used. For
productive systems an exception may be too devastating, but to make it
possible to find the issues nonetheless a warning has been added.

215665