ServletFilterHelper: reject unauthorized POST requests
It is possible for the UI server to receive a POST request without a
valid HTTP session. If the call expected JSON, the ServletFilterHelper
already returned a special "session timeout" JSON message. However, for
all other content-types, only a warning was logged and the request was
redirected to the login form.

This warning is misleading and returning the login.html form is most
likely not what the callers expects or can handle. Therefore, the
following changes were made for POST requests:
1. Log level reduced to INFO
2. Return HTTP status 403 (Forbidden) instead of forwarding the request
   to login.html.

One case of a non-JSON POST request is the "/unload" beacon sent by
Session.js when reloading the page.

215431

Change-Id: Ic2e6e9881883df77e5a57ff5fac5ff9e6a8c57f3
(cherry-picked from e7166b2a8390f14126ce8279aa3968f489602c25)
Reviewed-on: https://git.eclipse.org/r/111363
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>